Calculates a set of statistics for a given array of numbers.

calc.stat calculates a set of statistics for a given array of numbers.

Usage

calc.stat(x)

Arguments

x

a numeric array

Value

An array of calculated statistics of a given set of numbers

Details

calc.stat calculates a set of statistics for a given array of numbers. The calculated statistics are

Ntot = length of the array

Nunique = Number of unique elements

Min = Minimum value of the array

Max = Maximum value of the array

Mean = Mean value of the array

Median = Median value of the array

SD = Standard deviation value of the array

SE = Standard error value of the array

CVp = Percent coefficient of variation of the array

CI95u = Upper limit of the 95% confidence interval of the array

cI95l = Lower limit of the 95% confidence interval of the array

gMean = Geometric mean value of the array

gCVp = Geometric percent coefficient of variation of the array