Plots drug plasma concentration vs time data

dv.plot plots DV vs Time data.

Usage

dv.plot(
  df,
  xvar = "Time",
  yvar = "Conc",
  obsLog = FALSE,
  myXlab = "Time",
  myYlab = "Concentration",
  color = NULL,
  group = NULL,
  guide = TRUE,
  onlyLin = FALSE,
  onlyLog = FALSE,
  XYlog = FALSE,
  STRATY = ".",
  STRATX = ".",
  myYBr = waiver(),
  myXBr = waiver(),
  myYBrLog = waiver(),
  myXBrLog = waiver(),
  myYlim = NULL,
  myXlim = NULL,
  myYlimLog = NULL,
  myXlimLog = NULL,
  title = NULL
)

Arguments

df

A data frame to be used for the plot

xvar

is the independent variable, default is "TIME"

yvar

is the dependent variable, default is "DV"

obsLog

is a logical variable (TRUE, FALSE). If TRUE, concentration in observed data is assumed to be in logarithmic scale. Default is FALSE

myXlab

is the x-axis label, default is "Time"

myYlab

is the y-axis label, defaults is "Concentration"

color

is the column name of the color stratification variable, e.g. "DOSEF". Default is NULL

group

is the column name of the variable used to group data, default is "ID

guide

if TRUE, show guide, default is TRUE

onlyLin

if TRUE, presents only the linear version of the plot, default is FALSE

onlyLog

if TRUE, presents only the log version of the plot, default is FALSE

XYlog

if TRUE, both X and Y axes of the log version of the plot is shown on the logarithmic scale; if FALSE, only the Y-axis is shown on the logarithmic scale. Default is FALSE.

STRATY

is the row stratification variable, default is "."

STRATX

is the column stratification variable, default is "."

myYBr

are the breaks for the Y-axis for the linear plot

myXBr

are the breaks for the X-axis for the linear plot

myYBrLog

are the breaks for the Y-axis for the log plot

myXBrLog

are the breaks for the X-axis for the log plot

myYlim

sets Y-axis limits for the linear plot

myXlim

sets X-axis limits for the linear plot

myYlimLog

sets the Y-axis limit for the log plot

myXlimLog

sets the X-axis limit for the log plot

title

The title of the plot.

Value

returns a graphical object created by arrangeGrob function

Details

dv.plot plots DV vs Time data