Estimate individual NCA metrics.

Estimates a comprehensive set of NCA metrics for a given individual using concentration vs. time data.

Usage

est.nca(
  time,
  conc,
  backExtrp = FALSE,
  negConcExcl = FALSE,
  doseType = "ns",
  adminType = "extravascular",
  doseAmt = NULL,
  method = "linearup-logdown",
  AUCTimeRange = NULL,
  LambdaTimeRange = NULL,
  LambdaExclude = NULL,
  doseTime = doseTime,
  Tau = NULL,
  TI = NULL,
  simFile = NULL,
  dset = "obs",
  onlyNCA = FALSE,
  extrapolate = FALSE,
  sparse_compute = FALSE,
  force_extrapolate = FALSE,
  ...
)

Arguments

time

Numeric array for time

conc

Numeric array for concentration

backExtrp

If back-extrapolation is needed for AUC (TRUE or FALSE) (FALSE)

negConcExcl

Exclude -ve conc (FALSE)

doseType

Steady-state (ss) or non-steady-state (ns) dose ("ns")

adminType

Route of administration (iv-bolus,iv-infusion,extravascular) ("extravascular")

doseAmt

Dose amounts ("NULL")

method

Method to estimate AUC. The "linear" method applies the linear trapezoidal rule to estimate the area under the curve. The "log" method applies the logarithmic trapezoidal rule to estimate the area under the curve. The "linearup-logdown" method applies the linear trapezoidal rule to estimate the area under the curve for the ascending part of the curve and the logarithmic trapezoidal rule to estimate the area under the curve for the descending part of the curve.

AUCTimeRange

User-defined window of time used to estimate AUC ("NULL")

LambdaTimeRange

User-defined window of time to estimate elimination rate-constant ("NULL")

LambdaExclude

User-defined excluded observation time points for estimation of elimination rate-constant ("NULL")

doseTime

Dose time prior to the first observation for steady-state data (NULL)

Tau

Dosing interval for steady-state data ("NULL")

TI

Infusion duration ("NULL")

simFile

Name of the simulated concentration-time data if present ("NULL")

dset

Type, i.e., observed or simulated concentration-time data set ("obs" or "sim") ("obs")

onlyNCA

If TRUE only NCA is performed and ppc part is ignored although simFile is not NULL. Default is FALSE

extrapolate

Should the function extrapolate from the last observation to infinity?

sparse_compute

Should NCA metrics be computed even with only one sample?

force_extrapolate

Extrapolate AUC_inf with sparse data. Sparse data is defined as fewer than 3 points after Cmax for non-bolus input. In that case, Cmax is included in the extrapolation and only one extra point in required for extrapolation.

...

Arguments passed from other functions. Not used.

Value

An array of estimated NCA metrics

Details

est.nca estimates a comprehensive set of NCA metrics using the concentration-time profile of an individual. NCA metrics are estimated according to traditional PK calculations. The names of the various NCA metrics estimated in this package are assigned mainly following the names used in WinNonlin. This package accepts any of the three different types of drug administration, (i) iv-bolus, (ii) iv-infusion and (iii) extravascular; ncappc also can accept both non-steady state and steady-state data. The NCA metrics that are estimated and reported by ncappc are listed below.

C0 is the initial concentration at the dosing time. It is the observed concentration at the dosing time, if available. Otherwise it is approximated using the following rules.

Cmax, Tmax and Cmax_D are the value and the time of maximum observed concentration, respectively. If the maximum concentration is not unique, the first maximum is used. For steady state data, The maximum value between the dosing intervals is considered. Cmax_D is the dose normalized maximum observed concentration.

Clast and Tlast are the last measurable positive concentration and the corresponding time, respectively.

AUClast is the area under the concentration vs. time curve from the first observed to last measurable concentration.

AUMClast is the first moment of the concentration vs. time curve from the first observed to last measurable concentration.

MRTlast is the mean residence time from the first observed to last measurable concentration.

No_points_Lambda_z is the number of observed data points used to determine the best fitting regression line in the elimination phase.

AUC_pBack_Ext_obs is the percentage of AUCINF_obs that is contributed by the back extrapolation to estimate C0.

AUC_pBack_Ext_pred is the percentage of AUCINF_pred that is contributed by the back extrapolation to estimate C0.

AUClower_upper is the AUC under the concentration-time profile within the user-specified window of time provided as the "AUCTimeRange" argument. In case of empty "AUCTimeRange" argument, AUClower_upper is the same as AUClast.

Rsq, Rsq_adjusted and Corr_XY are regression coefficient of the regression line used to estimate the elimination rate constant, the adjusted value of Rsq and the square root of Rsq, respectively.

Lambda_z is the elimination rate constant estimated from the regression line representing the terminal phase of the concentration-time data.

Lambda_lower and Lambda_upper are the lower and upper limit of the time values from the concentration-time profile used to estimate Lambda_z, respectively, in case the "LambdaTimeRange" is used to specify the time range.

HL_Lambda_z is terminal half-life of the drug.

AUCINF_obs and AUCINF_obs_D are AUC estimated from the first sampled data extrapolated to infinity and its dose normalized version, respectively. The extrapolation in the terminal phase is based on the last observed concentration Clast_obs.

AUC_pExtrap_obs is the percentage of the AUCINF_obs that is contributed by the extrapolation from the last sampling time to infinity.

AUMCINF_obs is AUMC estimated from the first sampled data extrapolated to infinity. The extrapolation in the terminal phase is based on the last observed concentration.

AUMC_pExtrap_obs is the percentage of the AUMCINF_obs that is contributed by the extrapolation from the last sampling time to infinity.

Vz_obs is the volume of distribution estimated based on total AUC

Cl_obs is total body clearance.

AUCINF_pred and AUCINF_pred_D are AUC from the first sampled data extrapolated to infinity and its dose normalized version, respectively. The extrapolation in the terminal phase is based on the last predicted concentration obtained from the regression line used to estimate Lambda_z (Clast_pred).

AUC_pExtrap_pred is the percentage of the AUCINF_pred that is contributed by the extrapolation from the last sampling time to infinity.

AUMCINF_pred is AUMC estimated from the first sampled data extrapolated to infinity. The extrapolation in the terminal phase is based on the last predicted concentration obtained from the regression line used to estimate Lambda_z (Clast_pred).

AUMC_pExtrap_pred is the percentage of the AUMCINF_pred that is contributed by the extrapolation from the last sampling time to infinity.

Vz_pred is the volume of distribution estimated based on AUCINF_pred.

Cl_pred is the total body clearance estimated based on AUCINF_pred.

MRTINF_obs is the mean residence time from the first sampled time extrapolated to infinity based on the last observed concentration (Clast_obs).

MRTINF_pred is the mean residence time from the first sampled time extrapolated to infinity based on the last predicted concentration obtained from the regression line used to estimate Lambda_z (Clast_pred).

Tau is the dosing interval for steady-state data.

Cmin and Tmin are the minimum concentration between 0 and Tau and the corresponding time, respectively.

Cavg is the average concentration between 0 and Tau for steady-state data.

AUCtau and AUMCtau are AUC and AUMC between 0 and Tau for steady-state data.

Clss is an estimate of the total body clearance for steady-state data.

Vss_obs and Vss_pred are estimated volume of distribution at steady-state based on Clast_obs and Clast_pred, respectively.

p_Fluctuation is the percentage of the fluctuation of the concentration between 0 and Tau for steady-state data.

Accumulation_Index is \(1/(1-e^(-\lambda_z*\tau))\)