Plots population histogram of the NCA metrics selected for model diagnosis.

histpop.plot plots population histogram of the NCA metrics selected for model diagnosis (e.g. AUClast, AUCINF_obs, Cmax and Tmax).

Usage

histpop.plot(
  obsdata = outData,
  simdata = smedianData,
  figlbl = NULL,
  param = c("AUClast", "Cmax"),
  cunit = NULL,
  tunit = NULL,
  spread = "npi"
)

Arguments

obsdata

Data frame with the values of the NCA metrics estimated from the observed data

simdata

Data frame with the values of the NCA metrics estimated from the simulated data

figlbl

Figure label based on dose identifier and/or population stratifier (NULL)

param

A character array of the NCA metrics. The allowed NCA metrics for this histograms are "AUClast", "AUClower_upper", "AUCINF_obs", "AUCINF_pred", "AUMClast", "Cmax", "Tmax" and "HL_Lambda_z". (c("AUClast", "Cmax"))

cunit

Unit for concentration (default is NULL)

tunit

Unit for time (default is NULL)

spread

Measure of the spread of simulated data (ppi (95% parametric prediction interval) or npi (95% nonparametric prediction interval)) ("npi")

Value

returns a graphical object created by arrangeGrob function

Details

histpop.plot plots histogram of the NCA metrics selected for the model diagnosis and compares with the corresponding metrics estimated from the observed data. The allowed NCA metrics for this histograms are "AUClast", "AUClower_upper", "AUCINF_obs", "AUCINF_pred", "AUMClast", "Cmax", "Tmax" and "HL_Lambda_z". By default, this function produces histogram of AUClast and Cmax.