Check observed data

nca.check.obs Checks observed data for compatibility with ncappc and processes the data with various filtering criteria.

Usage

nca.check.obs(
  obsData,
  idNmObs = "ID",
  timeNmObs = "TIME",
  concNmObs = "DV",
  doseType = "ns",
  doseTime = NULL,
  Tau = NULL,
  filterNm = NULL,
  filterExcl = NULL,
  str1Nm = NULL,
  str1 = NULL,
  str2Nm = NULL,
  str2 = NULL,
  str3Nm = NULL,
  str3 = NULL,
  AUCTimeRange = NULL,
  LambdaTimeRange = NULL,
  adminType = "extravascular",
  TI = NULL,
  doseAmtNm = NULL,
  dateColNm = NULL,
  dateFormat = NULL,
  timeFormat = "number",
  concUnit = NULL,
  timeUnit = NULL,
  doseUnit = NULL,
  blqNm = NULL,
  blqExcl = 1,
  evid = TRUE,
  evidIncl = 0,
  mdv = FALSE
)

Arguments

obsData

Observed concentration-time data.

idNmObs

Column name for ID in observed data. Default is "ID"

timeNmObs

Column name for time in observed data. Default is "TIME"

concNmObs

Column name for concentration in observed data. Default is "DV"

doseType

Steady-state (ss) or non-steady-state (ns) dose. Default is "ns"

doseTime

Dose time prior to the first observation for steady-state data. Default is NULL

Tau

Dosing interval for steady-state data. Default is NULL

filterNm

Column name to filter data. Default is NULL

filterExcl

Row exclusion criteria based on the column defined by filterNm. This can be numeric value or logical condition (e.g. c(1, 2, "<20", ">=100", "!=100")). Default is NULL

str1Nm

Column name for 1st level population stratifier. Default is NULL

str1

Stratification ID of the members within 1st level stratification (e.g c(1,2)). Default is NULL

str2Nm

Column name for 2nd level population stratifier. Default is NULL

str2

Stratification ID of the members within 2nd level stratification (e.g c(1,2)). Default is NULL

str3Nm

Column name for 3rd level population stratifier. Default is NULL

str3

Stratification ID of the members within 3rd level stratification (e.g c(1,2)). Default is NULL

AUCTimeRange

User-defined window of time used to estimate AUC. Default is NULL

LambdaTimeRange

User-defined window of time to estimate elimination rate-constant. This argument lets the user to choose a specific window of time to be used to estimate the elimination rate constant (Lambda) in the elimination phase. The accepted format for the input to this argument is a numeric array of two elements; c(14,24) will estimate the Lambda using the data within the time units 14 to 24. Default is NULL

adminType

Route of administration. Allowed options are iv-bolus, iv-infusion or extravascular. Default is "extravascular"

TI

Infusion duration. If TI is a single numeric value, TI is the same for all individuals. If TI is the name of a column with numeric data present in the data set, TI is set to the unique value of the column for a given individual. Default is NULL

doseAmtNm

Column name to specify dose amount. Default is NULL

dateColNm

column name for date if used (e.g. "Date", "DATE"). Default is NULL

dateFormat

date format (D-M-Y, D/M/Y or any other combination of D,M,Y). Default is NULL

timeFormat

time format (number, H:M, H:M:S). Default is "number"

concUnit

Unit of concentration (e.g. "ng/mL"). Default is NULL

timeUnit

Unit of time (e.g. "h"). Default is NULL

doseUnit

Unit of dose amount (e.g. "ng"). Default is NULL

blqNm

Name of BLQ column if used to exclude data. Default is NULL

blqExcl

Excluded BLQ value; either a numeric value or a logical condition (e.g. 1 or ">=1" or c(1,">3")). Used only if the blqNm is not NULL. Default is "1"

evid

If TRUE EVID is used to filter data. Default is TRUE

evidIncl

Included values in EVID. Default is "0"

mdv

If TRUE MDV is used to include data when MDV=0. Default is FALSE

Value

A list of objects

Details

nca.check.obs Checks observed data for compatibility with ncappc.