Prepare individual PK data

nca.ind.data Extracts time-conc data for a given individual.

Usage

nca.ind.data(
  pkData,
  ID,
  dvLog = FALSE,
  dataType = "obs",
  idNm = "ID",
  timeNm = "TIME",
  concNm = "DV",
  adminType = "extravascular",
  TI = NULL,
  dateColNm = NULL,
  dateFormat = NULL,
  timeFormat = "number"
)

Arguments

pkData

PK concentration-time data.

ID

ID number of the individual.

dvLog

If TRUE concentration is in logarithmic scale. Default is FALSE

dataType

Indicates if the data is observed ("obs") or simulated ("sim"). Since the simulated data is assumed to be obtained from NONMEM output, DATE and clock time (dateColNm, dateFormat, timeFormat) are not used for time data. Default is "obs".

idNm

Column name for ID in PK data. Default is "ID"

timeNm

Column name for time in PK data. Default is "TIME"

concNm

Column name for concentration in PK data. Default is "DV"

adminType

Route of administration. Allowed options are iv-bolus, iv-infusion or extravascular. Default is "extravascular"

TI

Infusion duration. If TI is a single numeric value, TI is the same for all individuals. If TI is the name of a column with numeric data present in the data set, TI is set to the unique value of the column for a given individual. Default is NULL

dateColNm

column name for date if used (e.g. "Date", "DATE"). Default is NULL

dateFormat

date format (D-M-Y, D/M/Y or any other combination of D,M,Y). Default is NULL

timeFormat

time format (number, H:M, H:M:S). Default is "number"

Value

A list of objects with time-conc data and individual infusion duration for iv-infusion data

Details

nca.ind.data Extracts time-conc data for a given individual.