Plots population histogram of the NCA metrics selected for model diagnosis.

nca.npde.plot plots individual NPDE values and histogram of the NPDE values within a population group

Usage

nca.npde.plot(
  plotdata,
  xvar = NULL,
  npdecol = NULL,
  figlbl = NULL,
  cunit = NULL,
  tunit = NULL
)

Arguments

plotdata

Data frame with the values of the NPDE values of each individual for the NCA metrics

xvar

Name of the independent variable column against which NPDE values will be plotted

npdecol

Column names or column numbers of containing the NPDE values

figlbl

Figure label based on dose identifier and/or population stratifier (NULL)

cunit

Unit for concentration (default is NULL)

tunit

Unit for time (default is NULL)

Value

returns a data frame with the mean and SD of population NPDE values of each NCA metric and two graphical objects created by arrangeGrob function for the individual and population histogram of the NPDE values

Details

nca.npde.plot individual NPDE values and histogram of the NPDE values of NCA metrics within a population group.